debezium/dbz#2217 Map MySQL and MariaDB REAL columns to DOUBLE
MySQL and MariaDB treat REAL as a synonym for DOUBLE unless the
REAL_AS_FLOAT sql_mode is enabled; the server rewrites the column type
at DDL time and stores and replicates such columns as 8-byte doubles.

The ANTLR DDL parsers resolved the REAL keyword to java.sql.Types.REAL
(single precision in JDBC terms), so tables whose DDL was captured from
the binlog were emitted with a FLOAT32 schema and their values narrowed
via Float#floatValue(), silently losing precision. The snapshot path
parses the SHOW CREATE TABLE output, where the server already reports
the column as DOUBLE, producing a FLOAT64 schema - the same column
flipped between schemas depending on the capture path.

Resolve REAL to Types.DOUBLE in all three parsers so the streaming path
matches the snapshot path and no precision is lost. Default values for
REAL columns are now emitted as Double accordingly.

+++ debezium-connector-mysql/src/main/java/io/debezium/connector/mysql/antlr/MySqlAntlrDdlParser.java
@@ -190,7 +190,10 @@ public class MySqlAntlrDdlParser extends AntlrDdlParser<MySqlLexer, MySqlParser>
                 new DataTypeEntry(Types.BIGINT, MySqlParser.BIGINT_SYMBOL)
                         .setSuffixTokens(MySqlParser.SIGNED_SYMBOL, MySqlParser.UNSIGNED_SYMBOL, MySqlParser.ZEROFILL_SYMBOL),
                 // Floating point types
-                new DataTypeEntry(Types.REAL, MySqlParser.REAL_SYMBOL)
+                // REAL is by default a synonym for DOUBLE: the server stores and replicates such columns
+                // as 8-byte doubles. Only when the REAL_AS_FLOAT sql_mode is enabled is REAL a synonym
+                // for FLOAT instead (debezium/dbz#2217)
+                new DataTypeEntry(Types.DOUBLE, MySqlParser.REAL_SYMBOL)
                         .setSuffixTokens(MySqlParser.SIGNED_SYMBOL, MySqlParser.UNSIGNED_SYMBOL, MySqlParser.ZEROFILL_SYMBOL),
                 new DataTypeEntry(Types.DOUBLE, MySqlParser.DOUBLE_SYMBOL)
                         .setSuffixTokens(MySqlParser.PRECISION_SYMBOL, MySqlParser.SIGNED_SYMBOL, MySqlParser.UNSIGNED_SYMBOL, MySqlParser.ZEROFILL_SYMBOL),
